LayoutTests/accessibility/iframe-has-document.html

<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css" />
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

    <iframe id='iframe1' src='resources/cake.png'></iframe>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test makes sure the iframe's document is included in the main AX tree.");

    if (window.accessibilityController) {
        document.body.focus();
        var body = accessibilityController.focusedElement;
        var renderBlock = body.childAtIndex(0);
        var iframe = renderBlock.childAtIndex(0);
        var iframeDocument = iframe.childAtIndex(0);
        
        var parentIframeDocument = iframeDocument.parentElement();
        
        shouldBe("parentIframeDocument.isEqual(iframe)", "true");
    }

    successfullyParsed = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
