LayoutTests/accessibility/iframe-has-document-expected.txt


This test makes sure the iframe's document is included in the main AX tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parentIframeDocument.isEqual(iframe) is true
PASS successfullyParsed is true

TEST COMPLETE
